+++ applications/OptimizationApplication/python_scripts/responses/linear_strain_energy_response_function.py
@@ -13,8 +13,8 @@ import KratosMultiphysics.OptimizationApplication as KratosOA
 from KratosMultiphysics.OptimizationApplication.optimization_info import OptimizationInfo
 from KratosMultiphysics.OptimizationApplication.execution_policies.execution_policy_wrapper import ExecutionPolicyWrapper
 from KratosMultiphysics.OptimizationApplication.responses.response_function import ResponseFunction
-from KratosMultiphysics.OptimizationApplication.responses.response_function import ContainerEnum
-from KratosMultiphysics.OptimizationApplication.responses.response_function import GetSensitivityContainer
+from KratosMultiphysics.OptimizationApplication.utilities.helper_utils import ContainerEnum
+from KratosMultiphysics.OptimizationApplication.utilities.helper_utils import GetSensitivityContainer
 
 
 class LinearStrainEnergyResponseFunction(ResponseFunction):
