+++ applications/OptimizationApplication/tests/test_linear_strain_energy_response_function.py
@@ -7,7 +7,7 @@ import KratosMultiphysics.OptimizationApplication as KratosOA
 import KratosMultiphysics.KratosUnittest as kratos_unittest
 from KratosMultiphysics.kratos_utilities import DeleteFileIfExisting
 from KratosMultiphysics.OptimizationApplication.optimization_info import OptimizationInfo
-from KratosMultiphysics.OptimizationApplication.responses.response_function import ContainerEnum
+from KratosMultiphysics.OptimizationApplication.utilities.helper_utils import ContainerEnum
 from KratosMultiphysics.OptimizationApplication.execution_policies.execution_policy_wrapper import ExecutionPolicyWrapper
 from KratosMultiphysics.OptimizationApplication.responses.response_function_wrapper import ResponseFunctionBaseWrapper
 from KratosMultiphysics.OptimizationApplication.responses.response_function_wrapper import CreateResponseFunctionWrapper
